Changes to display auth token and fix the update for customer model
This commit adds the following fixes

1) The customer model edit dialog/form has been changed to show the REST
API authentication token. This field in the form is read only and is
displayed so that the user can use it for REST API communication.

2) There was a bug where the update action was not being acted upon. So
the fields in the customer model were not being committed to the
database. This has been fixed in this commit.