Fixing the issue where new user creation was not being created properly
When a customer is logged in, and adds a new user via the form, then the
user was being created in the database but the customer_id column was
not being populated with valid value. This happened because once we
started using friendly_id gem, the form gives back the friendly id
'slug' parameter instead of the id parameter. Now in this commit, the
proper customer_id is found and then assigned to the customer_id column
of the user object before commiting to the database.