Authenticating the customer before accessing the user views
When a customer object tries to access the user views, with this commit
we are enforcing that the customer is authenticated before being able to
access the views. This was causing a scenatio where - when a customer's
session times out and expires, and he refreshes the a user's url he was
able to access even though the html header shows that the user has been
signed out.