Displaying default values for some params for a new user form
When the customer logs in, and tries to create a new user, then this
commit adds default values for some parameters like max allocations,
default lifetime, max concurrent allocs and max bandwidth for the user.

Also in this commit, the first letter of the first name and last name
fields are capitalized, as has been done for customer model.