Adding field to db for webapp URL for CORS ACL check
In order to support checking for Access Control List checks, added a
webapp_url field to the Customer model. This can be used to check the
ORIGIN header field on the preflight request and reject if not from
provisioned URL. Checking is yet to be done, and will be added in the
next commit.