Adding validation of URL in the edit form for webapp URL field
When the user configures his web application URL after logging in, there
was no validation of the submitted URL. Hence now added validation of
the field value using the rails form helper 'url_field'. This does a
basic validation of the input field.
However, going ahead we need to add URL validation in the model and the
validation needs to be much accurate. This will be a todo, but the input
form checking done here will help for now.