Fixing the db constraint issue due to which signup is failing
Currently sign up is failing because we have a constrain in the customer
model that the webapp_url attribute/column can not be empty. However
during signing up, this value is not taken from the end user. Hence
removed the index for this column and removed the non-null constraint
using migration.