Migrating devise gem version from 2 from version 3
Migrating devise gem version from 2 from version 3. This is a somewhat
significant change because devise behavior has changed between ver 2 & 3
mainly around security. However, for our app the significant change has
been the removal of 'token_authenticatable' feature which we were using
for our REST API. The necessary functionality for replacing that feature
has been added in this commit in addition to other changes that were
necessiated by the migration to new version of the devise gem.

Some references used for the upgrade of devise from 2 to 3.

ref 1 - https://gist.github.com/josevalim/fb706b1e933ef01e4fb6
ref 2 - http://stackoverflow.com/questions/18605294/is-devises-token-authenticatable-secure
ref 3 - http://www.soryy.com/ruby/api/rails/authentication/2014/03/16/apis-with-devise.html
ref 4 - http://blog.plataformatec.com.br/2013/08/devise-3-1-now-with-more-secure-defaults/